Fix the logic surrounding the choice to record
The logic was not nearly as sane as it should have been. I was checking the
wrong thing and now it will record a response.

The next bump to overcome will be the fact that `json.load` returns unicode
data which is not compatible with my current RequestsBytesIO object. I'll have
to make that marginally more sophisticated to handle unicode content.