Allow to record duplicated urls on a cassette
Depending on the use case, it may be needed to record the
content of same url different times. Currently only one
url with the same match is allowed. This change proposes two
things:
1. If we are recording, do not query for matches, just record
everything that comes.

2. If we are consuming the cassette, consume each fixture once,
by setting a used flag. This will allow to consume each result
in order, and not duplicate.

Allow to record duplicated urls on a cassette

Depending on the use case, it may be needed to record the
content of same url different times. Currently only one
url with the same match is allowed. This change proposes two
things:
1. If we are recording, do not query for matches, just record
everything that comes.

2. If we are consuming the cassette, consume each fixture once,
by setting a used flag. This will allow to consume each result
in order, and not duplicate.

Fix tests and add new unit test for checking order

Fix tests for python 3.4