Update a Cassette's match_options attribute
Previously we were replacing the match_options attribute with either
the list that the user passes in, or the default list from
Cassette.default_cassette_options. If we return the latter, then users
who update the match_options attribute are updating the list instead
of an instance attribute.

Since each Cassette creates a new set, if we just update it, we avoid
this problem altogether.

Closes: gh-108