Handle differences in Requests 2.11.0
Requests 2.11.0 subtly changed some of the logic around requests
bodies - they will always be bytes objects on Python 3 - so we need
to update our body matcher logic to coerce the body unconditionally
if it exists for both the request from Requests and our recorded
Request.

Two identical bodies will be coerced identically so this should not
have backwards incompatible implications.

Closes #114